Move boolean logic in is own namespace
